added CURRENT_USER ANNOTATION to app/routers/security/dependencies

+++ app/routers/units.py
@@ -19,6 +19,8 @@ from app.services import units as unit_service
 
 from app.utils.mapper import map_string_to_model
 
+from app.routers.security.dependencies import CURRENT_USER
+
 
 router = APIRouter(prefix="/units", tags=["Units"])
 
@@ -26,34 +28,68 @@ UnitResponseModel = Annotated[units_schema.UnitOutput, Literal["UnitResponse"]]
 #✅
 @router.post("/", status_code=status.HTTP_201_CREATED, response_model=units_schema.UnitCreateOutput)
 def create_unit(
+    current_user: CURRENT_USER,
     unit: units_schema.UnitAdd, 
     vehicle: vehicles_schema.VehicleAdd, 
-    db: Session = Depends(get_db)) -> units_schema.UnitCreateOutput:
+    db: Session = Depends(get_db)
+    ) -> units_schema.UnitCreateOutput:
 
-    db_unit = unit_service.create_unit(db, unit=unit, vehicle=vehicle)
-
-    if db_unit is None:
-        raise HTTPException(status_code=404, detail="Unit not found")
+    if current_user.is_admin:
+        db_unit = unit_service.create_unit(db, unit=unit, vehicle=vehicle)
+        if not db_unit:
+            raise HTTPException(
+                status_code=400, 
+                detail=f"Unit with name {unit.name} already exists"
+                )
     return db_unit
 
+
+#✅
+@router.get("/{unit_id}", status_code=status.HTTP_200_OK, response_model=UnitResponseModel)
+def get_unit(
+    current_user: CURRENT_USER,
+    unit_id: int, 
+    db: Session = Depends(get_db),
+    include_vehicle: bool = False,
+    include_store: bool = False
+    ) -> UnitResponseModel:
+
+    if not current_user.is_active:
+        raise HTTPException(status_code=400, detail="Inactive user")
+
+    unit = unit_service.get_unit_by_id(
+        db, 
+        unit_id=unit_id,
+        include_vehicle=include_vehicle,
+        include_store=include_store
+        )
+
+    if not unit:
+        raise HTTPException(
+            status_code=404, 
+            detail=f"Unit with id {unit_id} not found"
+            )
+    print(f"CURRENT_USER : {CURRENT_USER}")
+    return unit
 
 
 #✅
 @router.get("/", status_code=status.HTTP_200_OK, response_model=List[UnitResponseModel])
 def get_units(
-              current_user: user_schema.UserInDB = Depends(get_current_active_user),  
-              db: Session = Depends(get_db),
-              skip: int = 0,
-              limit: int = 100,
-              filter_key: Optional[str] = None,
-              filter_value: Optional[str] = None,
-              to_join: bool = False,
-              models_to_join: Optional[Any] = None, # comma separated string of models to join
-              joined_model_filter_key: Optional[str] = None,
-              joined_model_filter_value: Optional[str] = None,
-              include_vehicle: bool = False,
-              include_store: bool = False,
-        ) -> UnitResponseModel:
+    current_user: CURRENT_USER,
+    db: Session = Depends(get_db),
+    skip: int = 0,
+    limit: int = 100,
+    filter_key: Optional[str] = None,
+    filter_value: Optional[str] = None,
+    to_join: bool = False,
+    models_to_join: Optional[Any] = None, # comma separated string of models to join
+    joined_model_filter_key: Optional[str] = None,
+    joined_model_filter_value: Optional[str] = None,
+    include_vehicle: bool = False,
+    include_store: bool = False,
+) -> List[UnitResponseModel]:
+
     if not current_user.is_active:
         raise HTTPException(status_code=400, detail="Inactive user")
     filter = {filter_key: filter_value} if filter_key and filter_value else None
@@ -84,35 +120,36 @@ def get_units(
     return units
 
 
-
-
-#✅
-@router.get("/{unit_id}", status_code=status.HTTP_200_OK, response_model=units_schema.UnitOutput)
-def get_unit(unit_id: int, db: Session = Depends(get_db)) -> units_schema.UnitOutput:
-    unit = unit_service.get_unit_by_id(db, unit_id=unit_id)
-    if not unit:
-        raise HTTPException(
-            status_code=404, 
-            detail=f"Unit with id {unit_id} not found"
-            )
-    return unit
-
-
 @router.delete("/{unit_id}", status_code=status.HTTP_200_OK, response_model=None)
-def delete_unit(unit_id: int, db: Session = Depends(get_db)) -> None:
-    unit_service.delete_unit(db, unit_id=unit_id)
-    return {
-        "Status": "Success", 
-        "Message": f"Unit with id {unit_id} deleted."
-    }
+def delete_unit(current_user: CURRENT_USER, unit_id: int, db: Session = Depends(get_db)) -> None:
+    if current_user.is_admin:
+        unit_service.delete_unit(db, unit_id=unit_id)
+        return {
+            "Status": "Success", 
+            "Message": f"Unit with id {unit_id} deleted."
+            }
+    else:
+        raise HTTPException(
+            status_code=403, 
+            detail=f"User {current_user.email} is not an admin"
+            )
+        
         
 @router.post("/expire_units", status_code=status.HTTP_200_OK)
 async def expire_units(background_tasks: BackgroundTasks, db: Session = Depends(get_db)) -> Any:
     background_tasks.add_task(unit_service.check_and_expire_units, db)
     return {"Status": "Success", "Message": "Expire units task added to background tasks."}
 
+
 @router.put("/{unit_id}", status_code=status.HTTP_200_OK, response_model=units_schema.UnitOutput)
-def update_unit(unit_id: int, unit: units_schema.UnitAdd, db: Session = Depends(get_db)) -> units_schema.UnitOutput:
+def update_unit(
+    current_user: CURRENT_USER,
+    unit_id: int, 
+    unit: units_schema.UnitAdd, 
+    db: Session = Depends(get_db)) -> units_schema.UnitOutput:
+
+    if not current_user.is_active:
+        raise HTTPException(status_code=400, detail="Inactive user")
     db_unit = unit_service.update_unit(db, unit=unit, unit_id=unit_id)
     if db_unit is None:
         raise HTTPException(status_code=404, detail="Unit not found")
